Changes:   - Only `/create` link present in header & removed connect wallet button   - Calculator page made index page (`/`) and previous index page moved to `/home`   - Disabled cover creator form submit button temporarily
