+++ components/CreateCoverForm/index.tsx
@@ -564,7 +564,8 @@ export const CreateCoverForm: FC = () => {
           text={`Create ${formData.coverName || "Cover"}`}
           type="submit"
           className="w-full mt-8 py-btn-y px-btn-x sm:px-16 lg:px-btn-x sm:w-auto"
-          disabled={submitDisabled}
+          // disabled={submitDisabled}
+          disabled
         />
       </form>
     </div>
